+++ bindings/python/orderflow/_ffi.py
@@ -823,6 +823,19 @@ class OrderflowLib:
         ]
         lib.of_get_metrics_json.restype = c_int32
 
+        lib.of_get_adapter_inventory_json.argtypes = [
+            ctypes.POINTER(c_char_p),
+            ctypes.POINTER(c_uint32),
+        ]
+        lib.of_get_adapter_inventory_json.restype = c_int32
+
+        lib.of_get_active_adapter_status_json.argtypes = [
+            c_void_p,
+            ctypes.POINTER(c_char_p),
+            ctypes.POINTER(c_uint32),
+        ]
+        lib.of_get_active_adapter_status_json.restype = c_int32
+
         lib.of_compute_lob_features.argtypes = [
             c_void_p,
             ctypes.POINTER(OfSymbol),